@@ -3,16 +3,16 @@ from unittest.mock import MagicMock, patch
 
 from requests import HTTPError, codes
 
-from src.client import players_season_totals
-from src.errors import InvalidSeason
-from src.http_service import HTTPService
+from src.scraper.api.client import players_season_totals
+from src.scraper.common.errors import InvalidSeason
+from src.scraper.services.http import HTTPService
 
 
 class TestPlayerSeasonTotals(TestCase):
     @patch.object(HTTPService, 'players_season_totals')
     def test_not_found_raises_invalid_season(self, mocked_players_season_totals):
         end_year = "jaebaebae"
-        expected_message = "Season end year of {end_year} is invalid".format(end_year=end_year)
+        expected_message = f"Season end year of {end_year} is invalid"
         mocked_players_season_totals.side_effect = HTTPError(response=MagicMock(status_code=codes.not_found))
         self.assertRaisesRegex(InvalidSeason, expected_message, players_season_totals, season_end_year=end_year)
 
