@@ -6,9 +6,9 @@ from unittest import TestCase
 
 import requests_mock
 
-from src.client import players_advanced_season_totals
-from src.data import OutputType, Team, Position
-from src.errors import InvalidSeason
+from src.scraper.api.client import players_advanced_season_totals
+from src.scraper.common.data import OutputType, Position, Team
+from src.scraper.common.errors import InvalidSeason
 
 
 class Test2019(TestCase):
@@ -16,7 +16,8 @@ class Test2019(TestCase):
         with open(os.path.join(
                 os.path.dirname(__file__),
                 "../files/player_advanced_season_totals/2019.html",
-        ), 'r') as file_input: self._html = file_input.read();
+        ), encoding="utf-8") as file_input:
+            self._html = file_input.read()
 
     @requests_mock.Mocker()
     def test_length(self, m):
@@ -38,7 +39,8 @@ class BaseTestPlayerAdvancedSeasonTotalsCSVOutput(TestCase):
         with open(os.path.join(
                 os.path.dirname(__file__),
                 f"../files/player_advanced_season_totals/{self.year}.html",
-        ), 'r') as file_input: self._html = file_input.read()
+        ), encoding="utf-8") as file_input:
+            self._html = file_input.read()
 
         self.output_file_path = os.path.join(
             os.path.dirname(__file__),
@@ -84,7 +86,8 @@ class BaseTestPlayerAdvancedSeasonTotalsJSONOutput(TestCase):
         with open(os.path.join(
                 os.path.dirname(__file__),
                 f"../files/player_advanced_season_totals/{self.year}.html",
-        ), 'r') as file_input: self._html = file_input.read()
+        ), encoding="utf-8") as file_input:
+            self._html = file_input.read()
 
         self.output_file_path = os.path.join(
             os.path.dirname(__file__),
@@ -111,8 +114,8 @@ class BaseTestPlayerAdvancedSeasonTotalsJSONOutput(TestCase):
             include_combined_values=self.include_combined_values,
         )
 
-        with open(self.output_file_path, "r", encoding="utf8") as output_file, \
-                open(self.expected_output_file_path, "r", encoding="utf8") as expected_output_file:
+        with open(self.output_file_path, encoding="utf-8") as output_file, \
+                open(self.expected_output_file_path, encoding="utf-8") as expected_output_file:
             self.assertEqual(
                 json.load(output_file),
                 json.load(expected_output_file),
@@ -254,8 +257,9 @@ class TestPlayerAdvancedSeasonTotalsInMemoryOutput(TestCase):
     def setUp(self):
         with open(os.path.join(
                 os.path.dirname(__file__),
-                f"../files/player_advanced_season_totals/2018.html",
-        ), 'r') as file_input: self._html = file_input.read()
+                "../files/player_advanced_season_totals/2018.html",
+        ), encoding="utf-8") as file_input:
+            self._html = file_input.read()
 
     def test_future_season_raises_invalid_season(self, m):
         m.get(f"https://www.basketball-reference.com/leagues/NBA_{sys.maxsize}_advanced.html",
@@ -266,7 +270,7 @@ class TestPlayerAdvancedSeasonTotalsInMemoryOutput(TestCase):
                                season_end_year=sys.maxsize)
 
     def test_2018_players_advanced_season_totals_length(self, m):
-        m.get(f"https://www.basketball-reference.com/leagues/NBA_2018_advanced.html",
+        m.get("https://www.basketball-reference.com/leagues/NBA_2018_advanced.html",
               text=self._html,
               status_code=200)
 
@@ -274,7 +278,7 @@ class TestPlayerAdvancedSeasonTotalsInMemoryOutput(TestCase):
         self.assertEqual(len(result), 605)
 
     def test_first_2018_players_advanced_season_totals_row(self, m):
-        m.get(f"https://www.basketball-reference.com/leagues/NBA_2018_advanced.html",
+        m.get("https://www.basketball-reference.com/leagues/NBA_2018_advanced.html",
               text=self._html,
               status_code=200)
 
@@ -316,7 +320,7 @@ class TestPlayerAdvancedSeasonTotalsInMemoryOutput(TestCase):
         )
 
     def test_last_2018_players_advanced_season_totals_row(self, m):
-        m.get(f"https://www.basketball-reference.com/leagues/NBA_2018_advanced.html",
+        m.get("https://www.basketball-reference.com/leagues/NBA_2018_advanced.html",
               text=self._html,
               status_code=200)
 
@@ -358,7 +362,7 @@ class TestPlayerAdvancedSeasonTotalsInMemoryOutput(TestCase):
         )
 
     def test_players_advanced_season_totals_json(self, m):
-        m.get(f"https://www.basketball-reference.com/leagues/NBA_2018_advanced.html",
+        m.get("https://www.basketball-reference.com/leagues/NBA_2018_advanced.html",
               text=self._html,
               status_code=200)
 
@@ -367,7 +371,7 @@ class TestPlayerAdvancedSeasonTotalsInMemoryOutput(TestCase):
             "./output/expected/player_advanced_season_totals/2018.json",
         )
         result = players_advanced_season_totals(season_end_year=2018, output_type=OutputType.JSON)
-        with open(expected_output_file_path, "r", encoding="utf8") as expected_output:
+        with open(expected_output_file_path, encoding="utf-8") as expected_output:
             self.assertEqual(
                 json.loads(result),
                 json.load(expected_output),
