@@ -3,7 +3,7 @@ from unittest import TestCase
 
 from lxml import html
 
-from src.html import DailyBoxScoresPage
+from src.scraper.html import DailyBoxScoresPage
 
 
 class TestDailyBoxScoresPage(TestCase):
@@ -11,7 +11,8 @@ class TestDailyBoxScoresPage(TestCase):
         with open(os.path.join(
                 os.path.dirname(__file__),
                 "../files/boxscores/2017/1/1.html"
-        ), 'r') as file_input: self.january_01_2017_box_scores = file_input.read()
+        ), encoding="utf-8") as file_input:
+            self.january_01_2017_box_scores = file_input.read()
 
     def test_game_url_paths_query(self):
         page = DailyBoxScoresPage(html=html.fromstring(self.january_01_2017_box_scores))
