@@ -3,11 +3,18 @@ from unittest import TestCase
 
 from lxml import html
 
-from src.data import Team, Position, POSITION_ABBREVIATIONS_TO_POSITION, \
-    TEAM_ABBREVIATIONS_TO_TEAM
-from src.html import PlayerSeasonTotalTable
-from src.parsers import PositionAbbreviationParser, TeamAbbreviationParser, \
-    PlayerSeasonTotalsParser
+from src.scraper.common.data import (
+    POSITION_ABBREVIATIONS_TO_POSITION,
+    TEAM_ABBREVIATIONS_TO_TEAM,
+    Position,
+    Team,
+)
+from src.scraper.html import PlayerSeasonTotalTable
+from src.scraper.parsers import (
+    PlayerSeasonTotalsParser,
+    PositionAbbreviationParser,
+    TeamAbbreviationParser,
+)
 
 
 class BasePlayerSeasonTotalsTestCase(TestCase):
@@ -18,7 +25,8 @@ class BasePlayerSeasonTotalsTestCase(TestCase):
         with open(os.path.join(
                 os.path.dirname(__file__),
                 f"../files/players_season_totals/{cls._season_end_year}.html",
-        ), 'r') as file_input: _html = file_input.read()
+        ), encoding="utf-8") as file_input:
+            _html = file_input.read()
         cls._parsed_season_totals = PlayerSeasonTotalsParser(
             position_abbreviation_parser=PositionAbbreviationParser(
                 abbreviations_to_positions=POSITION_ABBREVIATIONS_TO_POSITION
