@@ -2,10 +2,10 @@ from unittest import TestCase, mock
 
 from requests import HTTPError, codes
 
-from src.client import play_by_play
-from src.data import Team
-from src.errors import InvalidDate
-from src.http_service import HTTPService
+from src.scraper.api.client import play_by_play
+from src.scraper.common.data import Team
+from src.scraper.common.errors import InvalidDate
+from src.scraper.services.http import HTTPService
 
 
 class TestPlayByPlay(TestCase):
