@@ -2,9 +2,9 @@ from unittest import TestCase, mock
 
 from requests import HTTPError, codes
 
-from src.client import player_box_scores
-from src.errors import InvalidDate
-from src.http_service import HTTPService
+from src.scraper.api.client import player_box_scores
+from src.scraper.common.errors import InvalidDate
+from src.scraper.services.http import HTTPService
 
 
 class TestPlayerBoxScores(TestCase):
