@@ -1,11 +1,11 @@
 from unittest import TestCase
-from unittest.mock import patch, MagicMock
+from unittest.mock import MagicMock, patch
 
 from requests import HTTPError, codes
 
-from src.client import season_schedule
-from src.errors import InvalidSeason
-from src.http_service import HTTPService
+from src.scraper.api.client import season_schedule
+from src.scraper.common.errors import InvalidSeason
+from src.scraper.services.http import HTTPService
 
 
 class TestSeasonSchedule(TestCase):
