@@ -1,7 +1,7 @@
 from unittest import TestCase
 from unittest.mock import MagicMock
 
-from src.html import PlayerPage, PlayerPageTotalsTable
+from src.scraper.html import PlayerPage, PlayerPageTotalsTable
 
 
 class TestPlayerPage(TestCase):
