@@ -1,7 +1,7 @@
 from unittest import TestCase
-from unittest.mock import patch, PropertyMock, MagicMock
+from unittest.mock import MagicMock, PropertyMock, patch
 
-from src.html import PlayerSearchResult
+from src.scraper.html import PlayerSearchResult
 
 
 class TestPlayerSearchResult(TestCase):
