@@ -1,8 +1,8 @@
 from unittest import TestCase
 from unittest.mock import patch
 
-from src.data import PeriodType
-from src.parsers import PeriodDetailsParser
+from src.scraper.common.data import PeriodType
+from src.scraper.parsers import PeriodDetailsParser
 
 
 class TestPeriodDetailsParser(TestCase):
