@@ -1,7 +1,7 @@
 from unittest import TestCase
 
-from src.data import OUTCOME_ABBREVIATIONS_TO_OUTCOME, Outcome
-from src.parsers import OutcomeAbbreviationParser, PlayerBoxScoreOutcomeParser
+from src.scraper.common.data import OUTCOME_ABBREVIATIONS_TO_OUTCOME
+from src.scraper.parsers import OutcomeAbbreviationParser, PlayerBoxScoreOutcomeParser
 
 
 class TestPlayerBoxScoreOutcomeParser(TestCase):
