@@ -1,6 +1,6 @@
 from unittest import TestCase
 
-from src.parsers import SecondsPlayedParser
+from src.scraper.parsers import SecondsPlayedParser
 
 
 class TestSecondsPlayedParser(TestCase):
