@@ -2,17 +2,18 @@ from unittest import TestCase, mock
 
 from requests import codes
 
-from src.errors import InvalidDate
-from src.http_service import HTTPService
+from src.scraper.common.errors import InvalidDate
+from src.scraper.services.http import HTTPService
 
 
 class TestHTTPService(TestCase):
-    @mock.patch("requests.get")
-    def test_player_box_scores_raises_invalid_date_for_300_response(self, mocked_get):
+    def test_player_box_scores_raises_invalid_date_for_300_response(self):
+        service = HTTPService(parser=mock.MagicMock())
         response = mock.Mock(status_code=codes.multiple_choices)
-        mocked_get.return_value = response
-        self.assertRaisesRegex(
-            InvalidDate,
-            "Date with year set to 2018, month set to 1, and day set to 1 is invalid",
-            HTTPService(parser=mock.MagicMock()).player_box_scores,
-            day=1, month=1, year=2018)
+
+        with mock.patch.object(service, '_fetch', return_value=response):
+            self.assertRaisesRegex(
+                InvalidDate,
+                "Date with year set to 2018, month set to 1, and day set to 1 is invalid",
+                service.player_box_scores,
+                day=1, month=1, year=2018)
